Move all testing methods into testing module
