+++ packages/vm/src/instance.rs
@@ -229,8 +229,7 @@ where
 mod test {
     use super::*;
     use crate::errors::VmError;
-    use crate::mock::mock_dependencies;
-    use crate::testing::{mock_instance, mock_instance_with_balances};
+    use crate::testing::{mock_dependencies, mock_instance, mock_instance_with_balances};
     use crate::traits::ReadonlyStorage;
     use cosmwasm_std::{
         coin, from_binary, AllBalanceResponse, BalanceResponse, BankQuery, HumanAddr, Never,
@@ -558,11 +557,10 @@ mod test {
 #[cfg(test)]
 #[cfg(feature = "default-singlepass")]
 mod singlepass_test {
-    use crate::mock::mock_env;
     use cosmwasm_std::{coins, Never};
 
     use crate::calls::{call_handle, call_init, call_query};
-    use crate::testing::{mock_instance, mock_instance_with_gas_limit};
+    use crate::testing::{mock_env, mock_instance, mock_instance_with_gas_limit};
 
     static CONTRACT: &[u8] = include_bytes!("../testdata/contract.wasm");
