+++ packages/vm/src/testing/instance.rs
@@ -1,26 +1,19 @@
 //! This file has some helpers for integration tests.
 //! They should be imported via full path to ensure there is no confusion
 //! use cosmwasm_vm::testing::X
-use schemars::JsonSchema;
-use serde::{de::DeserializeOwned, Serialize};
+use cosmwasm_std::{Coin, HumanAddr};
 use std::collections::HashSet;
-use std::fmt;
 use std::iter::FromIterator;
 
-use crate::mock::{
-    mock_dependencies, mock_dependencies_with_balances, MockApi, MockQuerier, MockStorage,
-};
-use crate::{Api, Querier, Storage};
-use cosmwasm_std::{
-    to_vec, Coin, Env, HandleResult, HumanAddr, InitResult, QueryResponse, StdResult,
-};
-
-use crate::calls::{call_handle, call_init, call_query};
 use crate::compatability::check_wasm;
 use crate::instance::Instance;
+use crate::{Api, Querier, Storage};
+
+use super::mock::{mock_dependencies, mock_dependencies_with_balances, MockApi, MockQuerier};
+use super::storage::MockStorage;
 
 /// Gas limit for testing
-static DEFAULT_GAS_LIMIT: u64 = 500_000;
+const DEFAULT_GAS_LIMIT: u64 = 500_000;
 
 fn default_features() -> HashSet<String> {
     HashSet::from_iter(["staking".to_string()].iter().cloned())
@@ -52,58 +45,6 @@ pub fn mock_instance_with_gas_limit(
     check_wasm(wasm, &default_features()).unwrap();
     let deps = mock_dependencies(20, contract_balance);
     Instance::from_code(wasm, deps, gas_limit).unwrap()
-}
-
-// init mimicks the call signature of the smart contracts.
-// thus it moves env and msg rather than take them as reference.
-// this is inefficient here, but only used in test code
-pub fn init<
-    S: Storage + 'static,
-    A: Api + 'static,
-    Q: Querier + 'static,
-    T: Serialize + JsonSchema,
-    U: DeserializeOwned + Clone + PartialEq + JsonSchema + fmt::Debug,
->(
-    instance: &mut Instance<S, A, Q>,
-    env: Env,
-    msg: T,
-) -> InitResult<U> {
-    let serialized_msg = to_vec(&msg)?;
-    call_init(instance, &env, &serialized_msg).expect("VM error")
-}
-
-// handle mimicks the call signature of the smart contracts.
-// thus it moves env and msg rather than take them as reference.
-// this is inefficient here, but only used in test code
-pub fn handle<
-    S: Storage + 'static,
-    A: Api + 'static,
-    Q: Querier + 'static,
-    T: Serialize + JsonSchema,
-    U: DeserializeOwned + Clone + PartialEq + JsonSchema + fmt::Debug,
->(
-    instance: &mut Instance<S, A, Q>,
-    env: Env,
-    msg: T,
-) -> HandleResult<U> {
-    let serialized_msg = to_vec(&msg)?;
-    call_handle(instance, &env, &serialized_msg).expect("VM error")
-}
-
-// query mimicks the call signature of the smart contracts.
-// thus it moves env and msg rather than take them as reference.
-// this is inefficient here, but only used in test code
-pub fn query<
-    S: Storage + 'static,
-    A: Api + 'static,
-    Q: Querier + 'static,
-    T: Serialize + JsonSchema,
->(
-    instance: &mut Instance<S, A, Q>,
-    msg: T,
-) -> StdResult<QueryResponse> {
-    let serialized_msg = to_vec(&msg)?;
-    call_query(instance, &serialized_msg).expect("VM error")
 }
 
 /// Runs a series of IO tests, hammering especially on allocate and deallocate.
